test: extract common registration helpers to library
