+++ internal/testhelpers/config.go
@@ -1,6 +1,7 @@
 package testhelpers
 
 import (
+	"encoding/base64"
 	"testing"
 
 	"github.com/ory/kratos/driver/config"
@@ -22,5 +23,13 @@ func SetDefaultIdentitySchema(conf *config.Config, url string) {
 	conf.MustSet(config.ViperKeyDefaultIdentitySchemaID, "default")
 	conf.MustSet(config.ViperKeyIdentitySchemas, config.Schemas{
 		{ID: "default", URL: url},
+	})
+}
+
+// SetDefaultIdentitySchemaFromRaw allows setting the default identity schema from a raw JSON string.
+func SetDefaultIdentitySchemaFromRaw(conf *config.Config, schema []byte) {
+	conf.MustSet(config.ViperKeyDefaultIdentitySchemaID, "default")
+	conf.MustSet(config.ViperKeyIdentitySchemas, config.Schemas{
+		{ID: "default", URL: "base64://" + base64.URLEncoding.EncodeToString(schema)},
 	})
 }
