+++ internal/testhelpers/fake.go
@@ -1,0 +1,7 @@
+package testhelpers
+
+import "github.com/ory/x/randx"
+
+func RandomEmail() string {
+	return randx.MustString(16, randx.Alpha) + "@ory.sh"
+}
